% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_download.R
\name{aux_download}
\alias{aux_download}
\title{Download a single airbnb dataset}
\usage{
aux_download(link, dir = NULL, unzip = FALSE, summary = FALSE)
}
\arguments{
\item{link}{Link to a given airbnb dataset.}

\item{dir}{Directory to store downloaded data. If not specified, it will build
\code{airbnb_data/} directory.}

\item{unzip}{Boolean option. Whether to uzip saved datasets.
This opiton is only useful in case of raw data..}

\item{summary}{Boolean option. Add "summary_" prefix to file name
to distinguish raw dataset from summary dataset.}
}
\description{
Download a specific dataset from \url{http://insideairbnb.com} webpage.
}
\examples{
the_link <- "http://data.insideairbnb.com/the-netherlands/north-holland/amsterdam/2018-12-06/data/listings.csv.gz"
aux_download(the_link, unzip = TRUE)
}
\seealso{
\code{\link{download_airbnb_data}}
}
