% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_airbnb_data.R
\name{build_linkbase}
\alias{build_linkbase}
\title{Build linkbase}
\usage{
build_linkbase()
}
\value{
A data.frame with all information scraped from \url{http://insideairbnb.com} webpage.
}
\description{
Build a dataset based on scraped from \url{http://insideairbnb.com} webpage.
}
\keyword{internal}
