% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reviewed_flat.R
\name{calc_bayesian_mean}
\alias{calc_bayesian_mean}
\title{Calculate bayesian rating}
\usage{
calc_bayesian_mean(certain, prior, num_reviews, rating)
}
\arguments{
\item{certain}{A numeric value which defines the level of certainty.
It represents how confident the model is. It is equivalent to a number of observations.}

\item{prior}{A numeric value which defines mean of prior.}

\item{rating}{A list with rating for each flat.}

\item{num_review}{A list with number of reviews for each flat.}
}
\value{
A list with average bayesian rating
}
\description{
Calculate bayesian rating based on \url{https://fulmicoton.com/posts/bayesian_rating/}.
}
\keyword{internal}
