% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_airbnb_data.R
\name{extract_archived_info}
\alias{extract_archived_info}
\title{Extract data status}
\usage{
extract_archived_info()
}
\value{
A list with \code{archived} status.
}
\description{
Get information whether a given dataset is archived.
}
\keyword{internal}
