% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_download.R
\name{extract_file_name}
\alias{extract_file_name}
\title{Create file name from a link}
\usage{
extract_file_name(link, summary = F)
}
\arguments{
\item{link}{Link to a given airbnb dataset.}

\item{summary}{Boolean option. Add "summary_" prefix to file name
to distinguish raw dataset from summary dataset.}
}
\description{
Scrape webpage to create a proper file name for downloaded data.
}
\keyword{internal}
