% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_airbnb_data.R
\name{extract_links}
\alias{extract_links}
\title{Extract all links}
\usage{
extract_links()
}
\value{
A list with all links scraped from \url{http://insideairbnb.com} webpage.
}
\description{
Extract all available links with data from \url{http://insideairbnb.com} webpage.
}
\keyword{internal}
