% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Show a list with cities}
\usage{
extract_metadata(what)
}
\arguments{
\item{what}{Category you want to return. Available options: \code{"city"}
or \code{"country"}.}
}
\value{
A vector with all available cities.
}
\description{
Show all available cities to scrape from
the \url{http://insideairbnb.com} webpage.
}
\examples{
extract_metadata("city")
}
