% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_airbnb_data.R
\name{filter_linkbase}
\alias{filter_linkbase}
\title{Filter linkbase}
\usage{
filter_linkbase(data_to_load = NULL, archived = FALSE, city = NULL,
  format = NULL)
}
\arguments{
\item{data_to_load}{Names of datasets to download. Possible options:
`listings`, `calendar`, `reviews`. By default, it takes all of them into account.
A character scalar will save one dataset type, but by using vector it's possible
to define more dataset types.}

\item{archived}{Boolean option. Whether to download archived files as well.
By defualt,  \code{archived = FALSE} which means, it saves only the latest datasets.}

\item{city}{City name. The function will download only datasets related to the  \code{city}.
It could be character scalar or a vector with more cities. To see all available cities,
check \code{\link{extract_metadata}} function.}

\item{format}{Defined dataset format. Two options available:  \code{raw} which stores
raw data and  \code{summary} which have only the most important columns from raw data.
By default the function will download both formats.}
}
\value{
A data.frame with filtered links scraped from \url{http://insideairbnb.com} webpage.
}
\description{
Modify a scraped dataset based on filtering criteria.
}
\keyword{internal}
