% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_available_flat.R
\name{find_available_flat}
\alias{find_available_flat}
\title{Find available flat}
\usage{
find_available_flat(data, calendar, checkin, num_days)
}
\value{
A data.frame with flat which meet the available criterion.
}
\description{
Find only those flats which are available in a given period of time.
}
