% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_located_flat.R
\name{find_located_flat}
\alias{find_located_flat}
\title{Find located flat}
\usage{
find_located_flat(data, name, distance = NULL)
}
\arguments{
\item{name}{An assigned name for the distance column.}
}
\value{
A data.frame with flat which meet the distance criterion.
}
\description{
Modify a \code{listings} data.frame by defining column with distance
from city center. Additionally, filter out in terms of the distance from city center.
}
\examples{
\dontrun{
 find_located_flat(listings, distance_center, c(1,3)) \%>\%
    ggplot(., aes(xxx)) +
    geom_density()
}
}
