% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reviewed_flat.R
\name{find_reviewed_flat}
\alias{find_reviewed_flat}
\title{Find reviewed flat}
\usage{
find_reviewed_flat(data, name, certain = 30, prior = 90)
}
\arguments{
\item{data}{A data.frame with listings. If file was downloaded by
\code{download_airbnb_data} the name is "*listings.csv".}

\item{certain}{A numeric value which defines the level of certainty.
It represents how confident the model is. It is equivalent to a number of observations.}

\item{prior}{A numeric value which defines mean of prior.}
}
\value{
A data.frame with bayesian rating column.
}
\description{
Modify a \code{listings} data.frame by defining
average bayesian rating for each flat.
}
\examples{
\dontrun{
listings \%>\%
    find_reviewed_flat(bayes_review, certain = 30, prior = 95) \%>\%
    ggplot(aes(bayes_review)) + geom_histogram() +
    scale_x_continuous(breaks = seq(85, 100, 5))
}
}
