% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_your_place.R
\name{show_your_place}
\alias{show_your_place}
\title{Show your potential flats}
\usage{
show_your_place(data, top = 3)
}
\arguments{
\item{data}{A data.frame. Output of \code{\link{find_your_place}} function.}

\item{top}{Define top n flats to show on a map. The function will not show a map
if \code{top} is greater than 30.}
}
\value{
A leaflet map with your top n flats.
}
\description{
Produce a leaflet map to show top selected flats.
}
\examples{
\dontrun{
data_df <- find_your_place(listings, calendar, checkin = "2019-01-30",
                          num_days = 2, distance = c(1,3), num_guests = 2,
                          price = c(20, 100))
show_your_place(data_df)
}
}
