% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_priced_flat.R
\name{find_priced_flat}
\alias{find_priced_flat}
\title{Find priced flat}
\usage{
find_priced_flat(data, num_guests = NULL, rooms = NULL, price = NULL)
}
\value{
A data.frame with flat which meet all criteria.
}
\description{
Filter a \code{listings} data.frame in terms of some criteria.
}
\examples{
\dontrun{
find_priced_flat(listings,
                 num_guests = 3,
                 price = c(100, 200)) \%>\%
 ggplot(aes(price)) +
 geom_histogram()
}
}
