% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reviewed_flat.R
\name{find_reviewed_flat}
\alias{find_reviewed_flat}
\title{Find reviewed flat}
\usage{
find_reviewed_flat(data, name, certain = 30, prior = 90)
}
\value{
A data.frame with bayesian rating column.
}
\description{
Modify a \code{listings} data.frame by defining
average bayesian rating for each flat.
}
\examples{
\dontrun{
listings \%>\%
    find_reviewed_flat(bayes_review, certain = 30, prior = 95) \%>\%
    ggplot(aes(bayes_review)) + geom_histogram() +
    scale_x_continuous(breaks = seq(85, 100, 5))
}
}
